pyproject.toml
[tool.poetry]
name = "sns-sdk-python"
version = "0.1.0"
description = ""
authors = ["Clément Coutet <clement.coutet@grenoble-inp.org>"]
readme = "README.md"
packages = [
    { include = "src", from = "python" }
]
include = ["python/src/py.typed"]

[tool.poetry.dependencies]
python = "^3.10"
solders = "0.21.0"
solana = "^0.35.1"
pytest = "^8.3.3"
borsh-construct = "^0.1.0"
pytest-asyncio = "^0.24.0"
ruff = "^0.7.4"
mypy = "^1.13.0"
pynacl = "^1.5.0"
base58 = "^2.1.1"
bech32 = "^1.2.0"
python-dotenv = "^1.0.1"

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["python/src"]
testpaths = ["python/tests"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = "3.10"  # Specify your Python version
ignore_missing_imports = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
warn_unreachable = true
strict_optional = true

# Specify your source and test directories
files = ["python/src"]

